Fix bug related to the return type of str.find() function
The str.find() function returns a size_t type.  On 64-bits systems, like
amd64, storing the result of str.find() in an unsigned int variable is
bound to give problems.


git-svn-id: svn://svn.code.sf.net/p/octave/code/trunk/octave-forge@5866 fc592f45-0881-4044-91ff-c96495a54512

+++ main/miscellaneous/src/cell2csv.cc
@@ -92,7 +92,7 @@ DEFUN_DLD (cell2csv, args, nargout,
 	/* Output string value */
 	std::string str = c(i, j).string_value();
 	if (str.find(sep) != str.npos) {
-	  unsigned int pos = 0;
+	  size_t pos = 0;
 	  while ((pos=str.find(prot, pos)) != str.npos) {
 	    str.replace(pos, 1, prot+prot);
 	    pos += 2;
